NeverSink's itemfilter - version 4.4 economy/polishing update
CHANGES:

- Retiered all divination cards to better fit the current economy
- Retiered all uniques to better fit the current economy
- Increased the itemsizes of several low-mid tier orbs to fit the breach
pinatas and the clearspeed meta
- Alterations, chance, jeweler and chromatic orbs now have more
highlight ( https://imgur.com/a/0f1D6 )
- Added all corrupted belts to the show-list (amulets already there).
- You can now again click alt to check the ground for warband items (you
don't need to do it in the T13/T15 maps)

BUGFIXES:

- Fixed the highlight for finished "Breachstone" items. The official
information is that they are "Map Fragments". They are not. This is
client-side a bug. It's not major, but worth investigating to prevent
future problems /u/GGG_Bex
- ilvl75+ Spike point quivers now have the correct textcolor
- ilvl75+ 6S armors now have the correct textcolor
- Restructured a tiny bit - leather belts (for chancing) are now visible