NeverSink's itemfilter - v4.6 legacy version, improved map system, several improvements
- Retiered and repriced all uniques.
- Retiered divination cards and added 2 new cards to the tierlists
- Adjusted the leveling section a lot. It now has better highlights for
4L, useful racing gear, some flasks, "brutus-killing-equip" etc.
- Adjusted the chance bases
- Completely reworte the map tier lists. The new map tier list shows low
tier maps, and highlights them, WHEN you are in lowtier maps. Once you
outlevel the map drops (7 levels difference between the droptier and the
itemlevel) the map becomes "greyed out" and hidden on high strictness
version of the lootfilter. The exceptions are maps of tier 10 or higher
(never hidden) and all shaped maps.
- Adjusted several colors, settings, lot's and lot's of finetuning
- Gems with 14%+ qual are now more visible
- Lot's of minor adjustments to the structure, that improve the
performance and allow me to generate the filters easier.
- Fixed the blue/slick styles, sometimes still having some "old" green
backgrounds.
- Added a new "streamsound" style, that has an alt. soundsscheme - it's
designed for streamers and streamviewers to not get confused when the
other side find a good item.
- Adjusted the rare tierlist towards the upcoming economy
- Adjusted the crafting bases towards the upcoming economy
- Made the semi-strict filter more strict towards 2-handed and bulky
bases.
- Some stylistic and performance improvements - updated my tools to use
"tierlists" that group together different levels, basetypes, classes for
easier editing