6.15 beta READ THE README
I'm currently am struck with a fever: This is an updated, yet not super
polished verison of my filter. If I'll get better, I will update the
filter once again before the start of the flashback league.

Please provide feedback if you'll use it, so I know what to improve.

You'll find the 6.15 beta in the 6.15 beta folder and on
www.beta.filterblade.xyz (notice the BETA)

Thanks to Tobnac for helping with.. well everything, right now.

~~NeverSink